reorganize and purge more stuff
