+++ classification/feature_generation/fishing_feature_generation_test.py
@@ -1,0 +1,74 @@
+# Copyright 2017 Google Inc. and Skytruth Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import os
+import csv
+import model
+import numpy as np
+import utility
+import tensorflow as tf
+from datetime import datetime
+
+
+
+
+
+class _FakeRandint(object):
+    def randint(start, stop=0, step=0):
+        return 0
+
+
+class PythonFixedTimeExtractTest(tf.test.TestCase):
+    def test_cropped_extract(self):
+        with self.test_session():
+            input_data = np.array([[1., 5.], [2., 4.], [3., 7.], [4., 9.],
+                                   [5., 3.], [6., 8.], [7., 2.], [8., 9.]])
+
+            expected_result = np.array([[1., 5.], [2., 4.], [3., 7.], [4., 9.],
+                                        [5., 3.], [6., 8.], [1., 5.],
+                                        [2., 4.]])
+
+            class FakeRandomState(object):
+                def randint(self, min, max):
+                    return 0
+
+            res = utility.np_array_random_fixed_time_extract(
+                FakeRandomState(), input_data, 5, 8, 50)
+
+            self.assertAllEqual(res, expected_result)
+
+    def test_uncropped_extract(self):
+        with self.test_session():
+            input_data = np.array([[1., 5., 6.], [2., 4., 4.], [3., 7., 9.],
+                                   [4., 9., 0.]])
+
+            res = utility.np_array_random_fixed_time_extract(
+                _FakeRandint(), input_data, 20, 4, 50)
+            self.assertAllEqual(res, input_data)
+
+    def test_uncropped_extract_pad(self):
+        with self.test_session():
+            input_data = np.array([[1., 5., 6.], [2., 4., 4.], [3., 7., 9.]])
+            expected_result = np.array([[1., 5., 6.], [2., 4., 4.],
+                                        [3., 7., 9.], [1., 5.,
+                                                       6.], [2., 4., 4.]])
+
+            res = utility.np_array_random_fixed_time_extract(
+                _FakeRandint(), input_data, 20, 5, 50)
+            self.assertAllEqual(res, expected_result)
+
+
+
+if __name__ == '__main__':
+    tf.test.main()
