+++ classification/models/README.md
@@ -1,5 +1,3 @@
-# Models
+# Production
 
-Subdirectories located in this directory hold nnet models.
-
-Models used in production are in the directory `production`.
+Neural net classification models used in production.
