--- classification/models/dev/README.md
@@ -1,3 +1,0 @@
-# Dev
-
-Contains experimental nnet files not yet ready for production.
