--- classification/models/prod/README.md
@@ -1,3 +1,0 @@
-# Production
-
-Neural net classification models used in production.
